+++ ConsoleApp/Models/Entity/Account.cs
@@ -1,14 +1,11 @@
-﻿
-namespace ConsoleApp.Models
+﻿namespace ConsoleApp.Models.Entity
 {
-    public class Account
+    public class Account : BaseEntity
     {
         public string Title { get; set; } = string.Empty;
         public string Url { get; set; } = string.Empty;
         public string Username { get; set; } = string.Empty;
         public string Password { get; set; } = string.Empty;
         public string Details { get; set; } = string.Empty;
-        public DateTime CreatedOn { get; set; }
-        public DateTime ModifiedOn { get; set; }
     }
 }