+++ ConsoleApp/Models/Entity/BaseEntity.cs
@@ -1,0 +1,9 @@
+﻿
+namespace ConsoleApp.Models.Entity
+{
+    public class BaseEntity
+    {
+        public DateTime CreatedOn { get; set; }
+        public DateTime ModifiedOn { get; set; }
+    }
+}