+++ ConsoleApp/Repository/Abstraction/IGenericRepo.cs
@@ -1,0 +1,10 @@
+﻿using ConsoleApp.Models.Entity;
+
+namespace ConsoleApp.Repository.Abstraction
+{
+    public interface IGenericRepo<T> where T : BaseEntity
+    {
+        public List<T> GetAll();
+        public void Add(T entity);
+    }
+}